Clase 3/flask_traces/requirements.txt
Flask==3.0.0
gunicorn==21.2.0

# Google Cloud Profiler
google-cloud-profiler==4.1.0

# OpenTelemetry para Cloud Trace
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-gcp-trace
opentelemetry-resourcedetector-gcp